requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
openai>=1.82.0
sentry-sdk[fastapi,openai]>=2.29.0
python-dotenv>=1.1.0
